LayoutTests/fast/css/font-calculated-value.html

<!DOCTYPE html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<div id="testDiv""></div>
<script>
description("Tests assigning a calculated value to 'font' CSS property.");

var testDiv = document.getElementById("testDiv");

shouldBeEmptyString("testDiv.style['font']");
evalAndLog("testDiv.style['font'] = 'italic small-caps calc(100 * 9) 12px arial'");
shouldBeEqualToString("testDiv.style['font']", "italic small-caps 900 12px arial");
shouldBeEqualToString("window.getComputedStyle(testDiv).getPropertyValue('font')", "italic small-caps 900 normal 12px/14px arial");

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
